Update the ISA requirement section
Some of the ISA requirement sections do not belong to a platform
specification and can move to the profile specification. The fence.i
recommendation belong to software section as it talks about a requirement
for userspace applications. While at it, provide a better clarification
for the above requirement as well. Thanks Andrew & Greg for reviewing
this section.

Update the ISA requirement section to reflect that.

Suggested-by: Greg Favor <gfavor@ventanamicro.com>
Signed-off-by: Atish Patra <atishp@rivosinc.com>